+++ CHANGELOG.md
@@ -9,6 +9,10 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Fixed crashes when shooting, if gunflash or gunshell objects are not present in a level.
 * Fixed Teleporter object.
 * Fixed Wraith objects not working correctly in flipped rooms.
+* Fixed caustics not rendered correctly if texture compression was enabled.
+
+### Lua API changes
+* Fixed `Moveable.GetJointPosition` not returning correct results if moveable is invisible or not rendered.
 
 ### Lua API changes
 * Added AlignToMoveable function to `Lara` to allow alignment with moveables.